+++ src/main/java/ch/icclab/cyclops/dashboard/services/iaas/cloudstack/bills/BillPDF.java
@@ -106,7 +106,7 @@ public class BillPDF extends ServerResource {
                 while (keys.hasNext()) {
                     String key = (String) keys.next();
                     JSONObject billItem = (JSONObject) billDetails.get(key);
-                    Long usage = billItem.getLong("usage");
+                    Double usage = billItem.getDouble("usage");
                     Double cost = billItem.getDouble("price");
                     //String unit = billItem.getString("unit");
                     Double discount = billItem.getDouble("discount");
